fix : 반올림으로 캐싱 충돌 해결

+++ src/main/java/likelion/madi/repository/WeatherRepository.java
@@ -13,7 +13,7 @@ public interface WeatherRepository extends JpaRepository<Weather, Long> {
     // 💡 단건 조회 시 중복 데이터로 인한 예외를 막기 위해 List로 반환하도록 수정
     List<Weather> findByTargetDateAndCityAndDistrict(LocalDate targetDate, String city, String district);
 
-    List<Weather> findByTargetDateAndLatitudeAndLongitude(LocalDate targetDate, Integer latitude, Integer longitude);
+    List<Weather> findByTargetDateAndLatitudeAndLongitude(LocalDate targetDate, Double latitude, Double longitude);
 
     boolean existsByTargetDateAndCityAndDistrict(LocalDate targetDate, String city, String district);
 }
